Хендлеры в отлеьные файлы, тесты на ошибку 500, порт константой
